backend/requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.4.0
django-filter==25.1
dj-database-url==2.3.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
PyJWT==2.10.1
drf-spectacular>=0.27.0
requests==2.32.3
